+++ mapjs/src/browser/dom-map-widget.js
@@ -59,8 +59,9 @@ $.fn.domMapWidget = function (activityLog, mapModel, touchEnabled, imageInsertCo
 			'meta+plus ctrl+plus z': 'scaleUp',
 			'meta+minus ctrl+minus shift+z': 'scaleDown',
 			'Esc 0 meta+0 ctrl+0': 'resetView',
+			'alt+o': 'handleKey_loadMap',
 			'alt+s': 'saveMap',
-		// 'Esc': 'cancelCurrentAction'
+			// 'Esc': 'cancelCurrentAction'
 		},
 		charEventHandlers = {
 			'[': 'activateChildren',
@@ -174,7 +175,7 @@ $.fn.domMapWidget = function (activityLog, mapModel, touchEnabled, imageInsertCo
 				if (actOnKeys) {
 					event.stopImmediatePropagation();
 					event.preventDefault();
-					mapModel[mappedFunction]('keyboard');
+					mapModel[mappedFunction]('keyboard', event);
 				}
 			});
 		});